Terrain types calculation, label and mapmode
